+++ spring-cloud-tencent-dependencies/pom.xml
@@ -242,7 +242,7 @@
 
 			<dependency>
 				<groupId>com.tencent.cloud</groupId>
-				<artifactId>spring-cloud-starter-tencent-kafka-lane-plugin</artifactId>
+				<artifactId>spring-cloud-starter-tencent-message-queue-plugin</artifactId>
 				<version>${revision}</version>
 			</dependency>
