+++ spring-cloud-tencent-plugin-starters/pom.xml
@@ -26,7 +26,7 @@
 		<module>spring-cloud-starter-tencent-multi-discovery-plugin</module>
 		<module>spring-cloud-starter-tencent-traffic-mirroring-plugin</module>
 		<module>spring-cloud-starter-tencent-fault-injection-plugin</module>
-		<module>spring-cloud-starter-tencent-kafka-lane-plugin</module>
+		<module>spring-cloud-starter-tencent-message-queue-plugin</module>
 	</modules>
 
 </project>
