+++ spring-cloud-tencent-plugin-starters/spring-cloud-starter-tencent-message-queue-plugin/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  com.tencent.cloud.plugin.lane.kafka.KafkaLaneAspectConfiguration
